New Exceptions which all extends our ExceptionMiddleware which handles all error logging
