+++ MdsExceptions/InvalidResourceDataException.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace MdsExceptions;
+
+/**
+ * Class InvalidResourceDataException
+ * @package MdsSupportingClasses\Exceptions
+ */
+class InvalidResourceDataException extends ExceptionMiddleware
+{
+	/**
+	 * InvalidResourceDataException constructor.
+	 * @param string $message
+	 */
+	public function __construct($message)
+	{
+		parent::__construct($message);
+	}
+}
